CLI now able to get name based on episode & season
Implemented basic usage of the TV DB API - enough
to get the tv show ID which is then used to get
the name of the episode. HTTPary is used plus Icebox
which enables caching of responses.

I Also fixed up a bug in the CLI acceptance test where
I forgot to specify the command line arguments.