Changed the way TvDb{Client,ShowInfo} work which also affected Cli
ShowInfo#show_info now accepts an options hash as an argument

Cli#run now only accepts the API KEY as an argument, due to the
fact that ShowInfo#new takes a TV Show as an argument.

This also influenced the unit specs, which I then decided to
remove because the "happy path" is already being tested in the
acceptance specs, so now the unit specs only test particular
conditions that aren't covered by the acceptance tests.

(Enhancements based on Gregory Brown's feedback from RMU)